requirements.txt
streamlit
PyMuPDF
pytesseract
Pillow
langchain
langchain-community
faiss-cpu
sentence-transformers
